+++ template-sample-app/template/src/config/initialStateSlice.js
@@ -23,8 +23,6 @@ export const oauthInitialState = {
     clientId: '0m4N2QdVnJ48', // sample-app oauth app (@carto public user)
     scopes: [
       'user:profile', // to load avatar photo
-      'datasets:metadata', // to list all your datasets,
-      'datasets:r:*', // to read any of your datasets
       'dataservices:geocoding', // to use geocoding through Data Services API
       'dataservices:isolines', // to launch isochrones or isodistances through Data Services API
     ],