+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## Not released
+- Remove datasets section [#175](https://github.com/CartoDB/carto-react-template/pull/175)
 - Client Side Widgets [#159](https://github.com/CartoDB/carto-react-template/pull/159)
 - Fix basic tab layout id broken on first load [#162](https://github.com/CartoDB/carto-react-template/issues/162)
 - Add sources and views selector in hygen layer generator [#165](https://github.com/CartoDB/carto-react-template/pull/165)